--- a/ci/incremental/build.cmd
+++ b/ci/incremental/build.cmd
@@ -1,7 +1,7 @@
 @rem https://github.com/numba/numba/blob/master/buildscripts/incremental/build.cmd
 
 @rem Build extensions
-python setup.py build_ext -q -i -j4
+python setup.py build_ext -q -i
 
 @rem Install pandas
 python -m pip install --no-build-isolation -e .