@WillAyd 2019-12-11
I think could also reduce the cythonize bottleneck if using the `nthreads` argument:

https://cython.readthedocs.io/en/latest/src/userguide/source_files_and_compilation.html#Cython.Build.cythonize

Locally that yielded this for a clean build:

```sh
69.07s user 3.75s system 334% cpu 21.783 total
```

So 21 seconds from scratch isn't bad; might help a lot to reduce time to run ASVs