@jbrockmendel 2019-12-11
that would be a nice speedup.  im not clear on what the difference is between nthreads vs "-j"