@jbrockmendel 2019-12-11
ok, it looks like passing `nthreads` to `cythonize` affects the cython->C compilation, and the "-j 4" affects the C->so compilation