@WillAyd 2019-12-11
I'd roughly say this saves about 4-5 minutes on CI where it works. Have to see what's going on with Linux 36

Also not sure if Windows will really support this, at least not without some upstream changes. Earlier attempts throw a warning about multiprocessing `freeze_support`; put on the back burner for now