@topper-123 2019-12-19
I'm not too familiar with building Cython and that traceback doesn't give many clues AFAICT.

If this is a Windows problem, maybe then only pass the parallel arguments to build cython if we're not on windows?